Move file functions to their own file
