+++ 32blit-sdl/System.cpp
@@ -3,6 +3,7 @@
 #include <random>
 #include "SDL.h"
 
+#include "File.hpp"
 #include "System.hpp"
 #include "32blit.hpp"
 #include "UserCode.hpp"
@@ -60,40 +61,6 @@ uint32_t blit_random() {
 	return random_distribution(random_generator);
 }
 
-// file IO
-static std::string basePath;
-static std::map<uint32_t, SDL_RWops *> open_files;
-static uint32_t current_file_handle = 0;
-
-int32_t open_file(std::string name) {
-  auto file = SDL_RWFromFile((basePath + name).c_str(), "rb");
-
-  if(file) {
-    current_file_handle++;
-    open_files[current_file_handle] = file;
-    return current_file_handle;
-  }
-
-  return -1;
-}
-
-int32_t read_file(uint32_t fh, uint32_t offset, uint32_t length, char *buffer) {
-  auto file = open_files[fh];
-
-  if(file && SDL_RWseek(file, offset, RW_SEEK_SET) != -1) {
-    size_t bytes_read = SDL_RWread(file, buffer, 1, length);
-
-    if(bytes_read > 0)
-      return bytes_read;
-  }
-
-  return -1;
-}
-
-int32_t close_file(uint32_t fh) {
-  return SDL_RWclose(open_files[fh]) == 0 ? 0 : -1;
-}
-
 // SDL events
 const Uint32 System::timer_event = SDL_RegisterEvents(2);
 const Uint32 System::loop_event = System::timer_event + 1;
@@ -141,9 +108,7 @@ void System::run() {
 	blit::update = ::update;
 	blit::render = ::render;
 
-	auto basePathPtr = SDL_GetBasePath();
-	basePath = std::string(basePathPtr);
-	SDL_free(basePathPtr);
+	setup_base_path();
 
 	blit::open_file = ::open_file;
 	blit::read_file = ::read_file;
