+++ vs/32blit-sdl/32blit-sdl.vcxproj
@@ -202,6 +202,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="..\..\32blit-sdl\Audio.hpp" />
+    <ClInclude Include="..\..\32blit-sdl\File.hpp" />
     <ClInclude Include="..\..\32blit-sdl\Input.hpp" />
     <ClInclude Include="..\..\32blit-sdl\Renderer.hpp" />
     <ClInclude Include="..\..\32blit-sdl\System.hpp" />
@@ -209,6 +210,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\32blit-sdl\Audio.cpp" />
+    <ClCompile Include="..\..\32blit-sdl\File.cpp" />
     <ClCompile Include="..\..\32blit-sdl\Input.cpp" />
     <ClCompile Include="..\..\32blit-sdl\Main.cpp" />
     <ClCompile Include="..\..\32blit-sdl\Renderer.cpp" />
