+++ vs/32blit-sdl/32blit-sdl.vcxproj.filters
@@ -30,6 +30,9 @@
     <ClInclude Include="..\..\32blit-sdl\Audio.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\32blit-sdl\File.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\32blit-sdl\Input.cpp">
@@ -45,6 +48,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\..\32blit-sdl\Audio.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\32blit-sdl\File.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
